feat(calculator): Make clear button functional
Render AC button functional for clearing the calculator's input.
Accomplish this by generating the AC button separately from other
operator buttons and adding an event listener for a new function, clear,
which clears the calculator's input. Add comments in the function
generateCalculatorUI to delineate the generation of each type of button
(operator buttons, clear button, digit buttons) now that the function is
longer.